[predict] Change metric to write Learn.csv entries
When in deserialize mode (driver messages read from a file instead of
the minifilter), a metric is used to decide when Learn.csv lines
(used for model training) have to be written, in order to not overload
our model training with duplicates.

We think the number of driver messages read is a good metric to achieve
this goal. This, a training line is written every 50 driver messages
processed.