[predict] Add a path clustering library
This commit adds the slc-paths library which allows to cluster paths listed in a file.

The clustering is done from the distance between 2 paths.
We cluster by step the closest paths. If the distance becomes too important, we return the list of remaining clusters.
Clusters allow us to monitor the execution perimeter of a program.

You will also find 2 text files used to test the library.