[predict] Using the path clustering library in prediction
Clustering is done on folder paths where we have detected updates, creations or deletions.
We added to the process records the columns nb_clusters (number of clusters found) and clusters_max_size (the size of the biggest cluster).
We also changed the prediction condition, which are no longer based on the execution time between 2 records.