fix #20: high processor usage
Single-linkage clustering applied to the list of directories with modified files is used to feed `clusters` and `clusters_max_size` characteristics.

Considering average-linkage complexity is O(n<sup>2</sup>) and n may be big, clustering is computed on a dedicated thread with a frequency that depends on the last execution duration.