feat: better XGBOOST model
As we are more and more confident in our ability to enhance our models, we're delivering a new version which has been partially trained on synthetic data (SMOTE).

Also, we lower some artificial thresholds that were previously used to prevent false positive.

More generally speaking, we have to think about a better way to deliver and upgrade models on a regular basis.